feat(loan form): form can pre-populate fields and be used for editing existing loans; New loans can be submitted and appear in firestore

+++ lib/models/loan.dart
@@ -8,9 +8,9 @@ class Loan {
   String borrowerUsername;
   String financialPlatform;
   String borrowerName;
-  int amount;
-  int repayAmount;
-  int amountRepaid;
+  double amount;
+  double repayAmount;
+  double amountRepaid;
   Timestamp originationDate;
   Timestamp repayDate;
   String requestLink;
